adds tests for date  utils functions
